Optimise tokenisation and whitespace skipping
Tokenisation and whitespace skipping are two major hot spots for the
parser.

Optimise by:

1. Using re.exec() instead of str.replace()
2. Using re.lastIndex with sticky expressions to avoid repeatedly
   trimming the front of the string
3. Using str.indexOf('\n') when counting newlines
4. Using a zero-width lookahead to quickly reject integers in the "float" regexp
5. Tweaking other regexps
6. Peeking at the next character in tokenise() to reduce the number of
   regular expressions that need to be checked at each position
7. Moving initialisation of re tables out of the hot path
8. Also optimise the removal of quotation marks from strings by using
   str.substring() instead of str.replace().

Due to optimisation of the "whitespace" regexp tokenise() now splits
"  // foo" into two tokens (" " and "// foo") rather than one. This makes
no difference to the output of the parser. Verified with the test suite
and with the w-p-t "idlharness" tests.

With these changes tokenize() is 30% faster and all_ws() is no longer a
major contributor to parse time.