Tokenise a solitary '/' correctly
There was a regression where a '/' was not correctly tokenised as
"other" when it was not part of a comment. Fix it.

Add a regression test that a single '/' in an IDL file fails to parse
correctly.

Also split the cases of whitespace and comments in the tokeniser to make
them a bit faster.